backpack-trading-bot/requirements.txt
requests>=2.28.0
pandas>=1.5.0
numpy>=1.21.0
websocket-client>=1.4.0
python-dotenv>=0.19.0
ed25519>=1.5
cryptography>=3.4.0
ta-lib>=0.4.25
plotly>=5.11.0
dash>=2.7.0
schedule>=1.1.0
